+++ src/main/java/me/pajic/rearm/util/CompatFlags.java
@@ -11,10 +11,11 @@ public class CompatFlags {
             "idwtialsimmoedm", "desc",
             "item-descriptions", "description"
     );
-    public static boolean TRIMICA_LOADED = ReArm.xplat().isModLoaded("trimica");
-    public static boolean HMI_LOADED = ReArm.xplat().isModLoaded("hold-my-items");
-    public static boolean ENCHDESC_MOD_LOADED = ENCHANTMENT_DESCRIPTION_MODS.keySet().stream().anyMatch(mod -> ReArm.xplat().isModLoaded(mod));
-	public static boolean APOTHEOSIS_LOADED = ReArm.xplat().isModLoaded("apothic_attributes");
+    public static final boolean TRIMICA_LOADED = ReArm.xplat().isModLoaded("trimica");
+    public static final boolean HMI_LOADED = ReArm.xplat().isModLoaded("hold-my-items");
+    public static final boolean ENCHDESC_MOD_LOADED = ENCHANTMENT_DESCRIPTION_MODS.keySet().stream().anyMatch(mod -> ReArm.xplat().isModLoaded(mod));
+	public static final boolean APOTHEOSIS_LOADED = ReArm.xplat().isModLoaded("apothic_attributes");
+	public static final boolean PENCHANT_LOADED = ReArm.xplat().isModLoaded("penchant");
 
     public static boolean armorRebalanceActive() {
         return ReArm.CONFIG.armor.armorRebalance.get() && !APOTHEOSIS_LOADED;